changed type to int8

+++ test_model.py
@@ -15,7 +15,7 @@ def test_model(filename, time):
 
     # Test the model on random input data.
     input_shape = input_details[0]['shape']
-    input_data = np.array(np.random.random_sample(input_shape), dtype=np.float32)
+    input_data = np.array(np.random.random_sample(input_shape), dtype=np.int8)
     interpreter.set_tensor(input_details[0]['index'], input_data)
 
     start_time = time.time()
